refactor: update namespace and imports for VersionComparison, adjust LESS and template for footer update notice
- Migrate `VersionComparisonInterface` to `Model\VersionCheck` namespace
- Simplify CSS for footer update notice, replacing modifier class with `.major` child selector
- Adjust exception handler namespaces for standardization
