+++ app/design/adminhtml/Magento/backend/Magento_Backend/web/css/source/module/_footer.less
@@ -57,13 +57,13 @@
     }
 }
 
-.distro-update-notice {
+.page-footer .distro-update-notice {
     margin-left: .8rem;
-}
 
-.distro-update-notice--major {
-    color: @color-tomato-brick;
-    font-weight: @font-weight__bold;
+    .major {
+        color: @color-tomato-brick;
+        font-weight: @font-weight__bold;
+    }
 }
 
 //  Copyright
